Mergeo main con my branch actual para actualizar mi branch con cambios en client_client

+++ server/movement.cpp
@@ -98,11 +98,8 @@ int32_t Movement::calculateDistance(Movement &other) {
 void Movement::setChase(Movement &other, int speed) {
     int32_t x_difference = other.getX() - this->getX();
     int32_t y_difference = other.getY() - this->getY();
-    int32_t distance = sqrt(pow(x_difference, 2) + pow(y_difference, 2));
+    double distance = sqrt(pow(x_difference, 2) + pow(y_difference, 2));
     if (distance == 0) return;
-    //this is not optimal, we should see the distance
-    //between the 8 possible directions and pick the best
-    //but for now this will do for testing.
     double x_difference_double = x_difference;
     double y_difference_double = y_difference;
     double normalized_x = x_difference_double / distance;
